requirements.txt
argcomplete==3.6.2
asgiref==3.9.1
boto3==1.40.6
botocore==1.40.6
certifi==2025.8.3
cfgv==3.4.0
cfn-flip==1.3.0
charset-normalizer==3.4.3
click==8.2.1
distlib==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
django-solo==2.4.0
django-storages==1.14.6
djangorestframework==3.16.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
drf-yasg==1.21.10
django-cors-headers==4.7.0
boto3==1.40.6
django-storages==1.14.6
python-dotenv==1.1.1
pillow==11.3.0
zappa==0.60.2
ruff==0.8.4
pre-commit
drf-yasg==1.21.10
durationpy==0.10
filelock==3.18.0
hjson==3.1.0
identify==2.6.13
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
pillow==11.3.0
pip-chill==1.0.3
placebo==0.9.0
platformdirs==4.3.8
pre_commit==4.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.3
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.33.1
Werkzeug==3.1.3
zappa==0.60.2
